add "config for user modules"
added baseUrl option in tsconfig to use absolute paths instead of
relative paths while importing modules that will be created in the
project. This way even if the path of a module changes we do not
have to change it in all the files importing the module but just changing
in the paths below the baseUrl will sufficient.