modify "userName and password will be runtime params"
userName and password will be global parameters that can be sent while
running the specs. To run the specs we use protractor config.js command
with which we also have to pass the params. We need to find another way
of sending the params as a part of npm test command and that can be
used while running protractor config.js command.